requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
celery==5.3.4
redis==5.0.1
Pillow>=10.0.0
python-decouple==3.8
django-filter==23.5
drf-yasg==1.21.7
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2
gunicorn==21.2.0 